Change curl text box from disabled to readonly
and make the cursor look like you're only supposed to click on it.

It's supposed to auto-select-all when focused
but that doesn't seem to work well on firefox,
and selecting the text manually can be just as difficult.
Using readonly instead makes the selection work much better
and works the same as chrome.

Readonly also seems more appropriate
as "disabled" normally makes an element seem unusable.