Use code tag inside pre to work around chrome bug
The previous commit that html-escaped the json response caused Chrome to not
display large json strings (like the pod example in gh-2).  If you scroll
horizontally very far a portion would appear on the right side, but the left was
invisible.  You couldn't see the whole string without editing it, which might
end up truncating it.

Since the examples on the hljs site use pre > code I tried that and it fixed the
issue on Chrome.

Additionally both Chrome and Firefox would produce a horizontal scrollbar for
long strings whereas with the code tag they now wrap.
That may be a matter of preference but the vertical scrollbar seems more
consistent and intuitive to me than the horizontal scrollbar.
We could add `white-space: pre` to restore that functionality.