Separate body validation from change to avoid duplication
Previously we were triggering a change event when the textarea changed,
but our listener for the change event recent the textarea
which can cause the cursor to move in IE.

So separate the logic so that keyup triggers a model change and only validate on
change.  Also save the method on the object to so the function is only created
once.